requirements.txt
autopep8==1.6.0
numpy==1.22.3
scipy==1.8.0
scikit-image==0.19.2
tensorflow==2.8.0
keras==2.8.0
Pillow==9.1.0
gensim==4.1.2
nltk==3.7
torch==1.11.0
transformers==4.18.0
pymongo==3.12.0